+++ src/main/kotlin/xyz/cssxsh/mirai/plugin/command/ArknightsGuardCommand.kt
@@ -14,7 +14,7 @@ object ArknightsGuardCommand : CompositeCommand(
     @SubCommand("detail", "详情")
     @Description("查看蹲饼详情")
     suspend fun CommandSenderOnMessage<*>.detail() = sendMessage {
-        "当前蹲饼状态${task}, 蹲饼间隔${GuardInterval}".toPlainText()
+        "当前蹲饼状态${fromEvent.subject.delegate in GuardContacts}, 蹲饼间隔${GuardInterval}".toPlainText()
     }
 
     @SubCommand("speed", "速度")
@@ -28,14 +28,14 @@ object ArknightsGuardCommand : CompositeCommand(
     @SubCommand("open", "打开")
     @Description("开启提醒")
     suspend fun CommandSenderOnMessage<*>.open() = sendMessage {
-        task = true
+        GuardContacts.add(fromEvent.subject.delegate)
         "蹲饼已打开".toPlainText()
     }
 
     @SubCommand("close", "关闭")
     @Description("关闭提醒")
     suspend fun CommandSenderOnMessage<*>.close() = sendMessage {
-        task = false
+        GuardContacts.remove(fromEvent.subject.delegate)
         "蹲饼已关闭".toPlainText()
     }
 }
